% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_local.R
\name{normal_W}
\alias{normal_W}
\title{Degree normalization of adjacency matrix}
\usage{
normal_W(graph, normalize_strategy, weighted = TRUE)
}
\arguments{
\item{graph}{igraph object}

\item{normalize_strategy}{strategy of normalization, one of ["Row", "Column", "Laplacian"]}

\item{weighted}{whether it's a weighted graph, default TRUE}
}
\value{
Normalized matrix
}
\description{
Degree normalization of adjacency matrix
}
